% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{create_empty_fte_df}
\alias{create_empty_fte_df}
\title{Create empty FTE data frame}
\usage{
create_empty_fte_df()
}
\value{
Empty data frame with FTE enrollment schema
}
\description{
Create empty FTE data frame
}
\keyword{internal}
