% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_fte_enrollment}
\alias{download_fte_enrollment}
\title{Download FTE enrollment PDF}
\usage{
download_fte_enrollment(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Data frame with grade-level enrollment by county
}
\description{
Downloads the FTE Enrollment by Grade PDF from WVDE and parses it.
}
\keyword{internal}
