% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_headcount_enrollment}
\alias{download_headcount_enrollment}
\title{Download headcount enrollment PDF}
\usage{
download_headcount_enrollment(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Data frame with total enrollment by county
}
\description{
Downloads the Headcount Enrollment PDF from WVDE and parses it.
}
\keyword{internal}
