% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_district_name}
\alias{format_district_name}
\title{Format county name as district name}
\usage{
format_district_name(county_name)
}
\arguments{
\item{county_name}{County name (e.g., "KANAWHA")}
}
\value{
District name (e.g., "KANAWHA COUNTY SCHOOLS")
}
\description{
Converts county name to standard district name format.
West Virginia districts are named "COUNTY COUNTY SCHOOLS".
}
\keyword{internal}
