% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_district_id}
\alias{generate_district_id}
\title{Generate district ID from county name}
\usage{
generate_district_id(county_name)
}
\arguments{
\item{county_name}{County name}
}
\value{
Character district ID
}
\description{
Creates a standardized district ID based on county name.
Uses a simple two-letter abbreviation scheme.
}
\keyword{internal}
