% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_available_years}
\alias{get_available_years}
\title{Get available years of West Virginia enrollment data}
\usage{
get_available_years()
}
\value{
Integer vector of available years (2014-2025)
}
\description{
Returns a vector of years for which enrollment data is available
from the West Virginia Department of Education (WVDE).
}
\details{
Data comes from WVDE School Finance Data PDFs which are published
annually. Historical data availability depends on which years
WVDE has made available on their website.
}
\examples{
get_available_years()
}
