% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{get_raw_directory}
\alias{get_raw_directory}
\title{Get raw school directory data from WVDE}
\usage{
get_raw_directory()
}
\value{
Raw data frame as downloaded from WVDE
}
\description{
Downloads the raw school directory CSV file from the West Virginia
Department of Education WVEIS portal.
}
\keyword{internal}
