% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_raw_enr}
\alias{get_raw_enr}
\title{Download raw enrollment data from WVDE}
\usage{
get_raw_enr(end_year)
}
\arguments{
\item{end_year}{School year end (e.g., 2024 for 2023-24 school year)}
}
\value{
Data frame with enrollment data by county district
}
\description{
Downloads county-level enrollment data from West Virginia Department of
Education School Finance Data PDFs.
}
\keyword{internal}
