% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_wv_counties}
\alias{get_wv_counties}
\title{Get list of West Virginia counties}
\usage{
get_wv_counties()
}
\value{
Character vector of county names (uppercase)
}
\description{
Returns a vector of all 55 West Virginia county names.
West Virginia has 55 counties, each with one county school district.
}
\keyword{internal}
