% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_wvde_pdf_url}
\alias{get_wvde_pdf_url}
\title{Get URL for WVDE enrollment PDF}
\usage{
get_wvde_pdf_url(end_year, type = "FTE")
}
\arguments{
\item{end_year}{School year end}

\item{type}{Either "FTE" or "Headcount"}
}
\value{
Character URL
}
\description{
Constructs the URL for downloading enrollment PDFs from WVDE.
WVDE uses two URL patterns depending on the year.
}
\keyword{internal}
