% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{merge_superintendent_data}
\alias{merge_superintendent_data}
\title{Merge superintendent data with school directory}
\usage{
merge_superintendent_data(directory_data, supt_data)
}
\arguments{
\item{directory_data}{Processed directory data}

\item{supt_data}{Superintendent data from get_superintendent_data()}
}
\value{
Directory data with superintendent columns added
}
\description{
Joins superintendent information to the school directory by county.
}
\keyword{internal}
